refactor: fix codesandbox structure
