+++ src/codesandbox/Vanilla.ts
@@ -2,22 +2,29 @@
  * Copyright (c) 2020-present NAVER Corp.
  * egjs projects are licensed under the MIT license
  */
-import { joinStrs } from "../utils";
 import { CodeSandboxTemplate } from "../types";
 
-export const DEFAULT_VANILLA_CODESANDBOX = (dependencies: string[]) => new Function(`
-var previews = arguments[0];
-return {
-    framework: "vanilla",
-    files: {
-        "index.html": {
-            template: "html",
-            html: previews["HTML"],
-            cssFiles: ["src/styles.css"],
-            jsFiles: ["src/index.ts"],
+
+export const DEFAULT_VANILLA_CODESANDBOX: CodeSandboxTemplate = (userDependencies = [], files = {}) => {
+    return {
+        template: "parcel",
+        files: {
+            "index.html": {
+                tab: "HTML",
+                template: "html",
+                values: {
+                    cssFiles: ["src/styles.css"],
+                    jsFiles: ["src/index.ts"],
+                },
+            },
+            "src/styles.css": {
+                tab: "CSS",
+            },
+            "src/index.ts": {
+                tab: "Vanilla",
+            },
+            ...files,
         },
-        "src/styles.css": previews["CSS"] ? previews["CSS"][0] : "",
-        "src/index.ts": previews["Vanilla"][0],
-    },
-    userDependencies: ${joinStrs(dependencies)},
-};`) as CodeSandboxTemplate;
+        userDependencies,
+    };
+};
