--- src/presets/Vanilla.ts
@@ -1,7 +1,0 @@
-/**
- * Copyright (c) 2020-present NAVER Corp.
- * egjs projects are licensed under the MIT license
- */
-export const VANILLA_PRESET = {
-    template: "parcel",
-};
