--- src/presets/index.ts
@@ -1,8 +1,0 @@
-export * from "./Angular";
-export * from "./React";
-export * from "./ReactJS";
-export * from "./Preact";
-export * from "./Lit";
-export * from "./Svelte";
-export * from "./Vanilla";
-export * from "./Vue";
